Mobile: Refactoring
