+++ src/Frontend/SampleEcommerce.Moblie/SampleEcommerce.Moblie/SampleEcommerce.Moblie/Services/Dialog/IDialogService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace SampleEcommerce.Mobile.Abstractions
+namespace SampleEcommerce.Mobile.Services.Dialog
 {
     public interface IDialogService
     {
